Main API

Types

LinearNoiseApproximation.LNASystem — Type

A type that contains the information of the LNA system.

Fields

rn::ReactionSystem: a reaction network that is used to construct the LNA system, which is from Catalyst.jl

odesys::ODESystem: the linear noise approximation system, it is a ODESystem that extends the ReactionSystem by adding the covariance terms

kwargs::Any: Optional keyword arguments

Functions

LinearNoiseApproximation.get_ode_propensity — Function

Function to obtain the ODEs according to the rate equations

Arguments

rn::ReactionSystem: a reaction network that is used to construct the LNA system, which is from Catalyst.jl

combinatoric_ratelaws::Bool: (Optional) whether to use the combinatoric rate laws, default is false

LinearNoiseApproximation.get_symbolic_matrix — Function

Function to obtain the symbolic matrix of the LNA system

Arguments

rn::ReactionSystem: a reaction network that is used to construct the LNA system, which is from Catalyst.jl

combinatoric_ratelaws::Bool: (Optional) whether to use the combinatoric rate laws, default is false

LinearNoiseApproximation.find_states_cov_number — Function

Function to find the index of the covariance matrix in the LNA system.

Arguments

num::Int: the index of the species in the reaction network

lna_sys::LNASystem: the LNA system

LinearNoiseApproximation.expand_initial_conditions — Function

Function to expand the initial conditions of the LNA system.

Arguments

sys::LNASystem: the LNA system

u0::Vector: the initial conditions of the orignal system (rate equations)